+++ src/main/java/kr/modusplant/api/signup/social/model/request/SocialLoginRequest.java
@@ -1,4 +1,4 @@
-package kr.modusplant.api.crud.model.request;
+package kr.modusplant.api.signup.social.model.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Getter;
