+++ src/main/java/kr/modusplant/api/signup/social/model/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package kr.modusplant.api.crud.model.response;
+package kr.modusplant.api.signup.social.model.response;
 
 import lombok.Getter;
 import lombok.Setter;
